requirements.txt
# HVAC GraphRAG Dependencies
# For Daniel @ Multipoint S.A.

# Core dependencies
neo4j==5.26.0
pyyaml==6.0.1
python-dotenv==1.0.0

# FastAPI and web framework (for future API)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# OpenAI and embeddings
openai==1.3.8
tiktoken==0.5.2

# Vector database
qdrant-client==1.7.0

# Utilities
requests==2.31.0
pandas==2.1.4
numpy==1.26.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
